+++ src/main/resources/static/css/admin/admin.css
@@ -348,7 +348,6 @@ main {
     background: #408D1C;
     border-radius: 50px;
     z-index: 10;
-
 }
