+++ src/main/resources/static/css/reset.css
@@ -102,3 +102,22 @@ nav img, nav span:nth-child(2), nav span:nth-child(3), nav span:nth-child(4), na
     /* 남호정, 소개, 공지사항, 게시판, mbti, 마이페이지, 로그인  */
     cursor : pointer;
 }
+
+#logoutNav{
+    /* 로그아웃 */
+    position: absolute;
+    width: 100px;
+    height: 21px;
+    left: 1670px;
+    top: 29px;
+    font-family: 'BMJUA';
+    font-weight: 400;
+    font-size: 20px;
+    line-height: 21px;
+    color: #C9E6CA;
+    transition: all 1s;
+    background: none;
+    border:none;
+    text-align: center;
+    cursor: pointer;
+}
