+++ src/main/resources/templates/user/mypage/userinfo.html
@@ -12,16 +12,16 @@
 
     <nav>
         <!--네비바-->
-        <img th:src="@{/images/main/nav-logo-dark.png}" alt="상단 로고" onclick="window.location.href='/'">
+        <img th:src="@{/images/main/nav-logo-dark.png}" alt="상단 로고" onclick="location.href='/'">
         <span onclick="location.href='/intro'">소개</span>
         <span onclick="location.href='/notice'">공지사항</span>
         <span onclick="location.href='/post'">게시판</span>
-        <span>MBTI 검사</span>
-        <span id="mypageBtn" style="color: #A3DA87;">마이페이지</span>
+        <span onclick="location.href='/mbti'">MBTI 검사</span>
+        <span id="mypageBtn">마이페이지</span>
         <form th:action="@{/auth/logout}" method="post">
-        <button id="logoutNav" type="submit">로그아웃</button>
+            <button id="logoutNav" type="submit">로그아웃</button>
         </form>
-        <span>회원가입</span>
+        <span onclick="location.href='/auth/signupterms'">회원가입</span>
     </nav>
 
 </header>
